% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcas_rcpp.R
\name{hcas_bench}
\alias{hcas_bench}
\title{Benchmarking}
\usage{
hcas_bench(
  observed,
  predicted,
  samples,
  histogram,
  xy_stats = c(0, 0, 1, 1),
  xy_penalty = 0,
  within_km = 200,
  k_env = 50,
  k_rs = 20,
  bin_width = 0.03,
  interpolate = TRUE,
  offset = 0,
  confidence = 0.5,
  lambda = 2,
  make_su = FALSE,
  force_compile = FALSE,
  num_threads = parallel::detectCores() - 1,
  filename = "",
  overwrite = TRUE,
  source_code
)
}
\arguments{
\item{source_code}{}
}
\description{
The HCAS benchmarking function.
}
