% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{hcas_color}
\alias{hcas_color}
\title{The HCAS colors}
\usage{
hcas_color(n = 10)
}
\arguments{
\item{n}{Integer. Number of color codes to return.}
}
\value{
a charterer vector of color codes
}
\description{
The HCAS colors
}
