% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{histo_color}
\alias{histo_color}
\title{The HCAS histogram colors}
\usage{
histo_color(n = 10)
}
\arguments{
\item{n}{Integer. Number of color codes to return.}
}
\value{
a charterer vector of color codes
}
\description{
The HCAS histogram colors
}
