% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo_rcpp.R
\name{histo_hcas}
\alias{histo_hcas}
\title{Histogram}
\usage{
histo_hcas(
  observed,
  predicted,
  samples,
  within_km = 1000,
  bin_width = 0.05,
  bin_num = 650,
  force_compile = FALSE,
  num_threads = parallel::detectCores() - 1,
  output_file,
  source_code
)
}
\arguments{
\item{source_code}{}
}
\description{
The HCAS histogram calcualtion based on pair-point densities.
}
