% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{The HCAS histogram}
\usage{
histogram(
  observed,
  predicted,
  samples,
  within_km = 1000,
  bin_width = 0.05,
  bin_num = 650,
  num_threads = parallel::detectCores() - 1,
  filename = "",
  source_code
)
}
\arguments{
\item{observed}{The data.frame or matrix of observed remote sensing variables.}

\item{predicted}{The data.frame or matrix of predicted remote sensing variables.}

\item{samples}{Matrix or df. With the XY or XY and RS and ENV values.}

\item{within_km}{Numeric. Search radius in kilometers considering samples.}

\item{bin_width}{Numeric. The bin width of the histogram}

\item{bin_num}{Int. Number of bin for histogram. Keep it at the default 650..}

\item{num_threads}{Int. Number of CPU threads for processing...}

\item{filename}{Char (optional). The output file name for the .text file.}
}
\value{
matrix
}
\description{
The HCAS histogram calculation is based on pair-point densities. It is an integral part of HCAS,
designed to learn the expected observed RS values from the predicted RS values across a wide range
of reference sites. Note that the reference sites used for the \code{histogram} do not need to be
the same as those used in the \code{\link{benchmark}} function. See more in details.
}
\details{
Ensure that the order of RS variables is consistent between observed and predicted inputs.
The RS variable values must be centered and scaled prior to prediction. Failure to do
so may result in variables with larger ranges having disproportionate influence in the
multi-dimensional distance calculations.

Ensure that \href{https://en.wikipedia.org/wiki/OpenMP}{OpenMP} is installed on your system
to take advantage of parallel processing and accelerate computations. While most systems
include OpenMP by default, you may need to load the appropriate module if you're using an HPC
system.
}
\examples{
\donttest{
library(ClassicHCAS)

obs_dat <- read.csv("")
prd_dat <- read.csv("")


}
}
\seealso{
\code{\link{benchmark}}
}
