% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ref_density_color}
\alias{ref_density_color}
\title{The HCAS reference density colors}
\usage{
ref_density_color(n = 10)
}
\arguments{
\item{n}{Integer. Number of color codes to return.}
}
\value{
a charterer vector of color codes
}
\description{
The HCAS reference density colors
}
