% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarking target points}
\usage{
benchmark(
  observed,
  predicted,
  samples,
  histogram,
  xy_stats = c(0, 0, 1, 1),
  xy_penalty = 0,
  within_km = 200,
  k_env = 50,
  k_rs = 20,
  bin_width = 0.05,
  interpolate = TRUE,
  offset = 0,
  confidence = 0.5,
  lambda = 2,
  exclude_slef = TRUE,
  make_su = FALSE,
  num_threads = parallel::detectCores() - 1,
  filename = "",
  overwrite = TRUE
)
}
\arguments{
\item{observed}{The data.frame, matrix or SpatRaster of observed remote sensing variables.}

\item{predicted}{The data.frame, matrix or SpatRaster of observed remote sensing variables.}

\item{samples}{Matrix or df. With the XY or XY and RS and ENV values.}

\item{histogram}{A matrix of HCAS histogram.}

\item{xy_stats}{Vector, mean and standard deviation of coordinates for centre and
scaling the coordinate to use as a penalty. The order should be: mean(x), mean(y), sd(x), sd(y)}

\item{xy_penalty}{Numeric. The penalty value the higher the value the more penalise the
distant location will be. The value 0 means no penalty}

\item{within_km}{Numeric. Search radius in kilometers for considering benchmark samples.}

\item{k_env}{Integer. Number of nearest ENV/predicted RS samples to take.}

\item{k_rs}{Integer. Number of nearest observed RS sample to takes.}

\item{bin_width}{Numeric. The bin width of the histogram}

\item{interpolate}{Logical. Should interpolate the histogram?}

\item{offset}{Int. Number of histogram bins to ignore...}

\item{confidence}{Numeric. The confidence value for LDC methods. See details below..}

\item{lambda}{Numeric. The lambda param for LDC Cauchy weighting...}

\item{make_su}{Logical. To make the uncertainty map or not.}

\item{num_threads}{Int. Number of CPU threads for processing...}

\item{filename}{Char (optional). The output file name for raster outputs}

\item{overwrite}{Logical. Whether to overwrite the output raster.}
}
\value{
matrix or SpatRaster
}
\description{
The HCAS benchmarking function to calculate habitat condition based on observed and
predicted remote sensing variables.
}
