% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarking target points}
\usage{
benchmark(
  data,
  samples,
  histogram,
  xy_stats = c(0, 0, 1, 1),
  xy_penalty = 0,
  radius_km = 200,
  scale_factor = 1e+05,
  k_pred = 50,
  k_obs = 20,
  bin_width = NULL,
  interpolate = TRUE,
  offset = 0,
  confidence = 0.5,
  lambda = 2,
  exclude_slef = TRUE,
  drop_features = NULL,
  make_su = FALSE,
  num_threads = -1,
  ...
)
}
\arguments{
\item{data}{A matrix, SpatRaster (from the \pkg{terra} package), or data.frame containing the input data.
The data must be organized in the following order: \strong{x}, \strong{y}, \strong{observed-RS},
\strong{predicted-RS} variables. If using a SpatRaster, the \strong{x} and \strong{y} are not required.
If using a matrix or data.frame, ensure that the variables are in the
correct order. For more see the details section.}

\item{samples}{A matrix or data.frame containing x, y, observed-RS, and predicted-RS values
(in that specific order) for benchmark samples (also known as reference sites). If the
\code{data} argument is a SpatRaster, you can provide only the x and y coordinates of the
benchmark samples. In this case, the corresponding values will be extracted from the raster
layers. Consider extra time for sample value extraction in this case.}

\item{histogram}{A matrix or \strong{histo} object of normalised HCAS histogram
see \code{\link{histogram}} and \code{\link{normalise}}).}

\item{xy_stats}{A vector, mean and standard deviation of coordinates for centre and
scaling the coordinate to use as a penalty. The order should be: mean(x), mean(y), sd(x), sd(y)}

\item{xy_penalty}{Numeric. The spatial distance penalty value for selecting benchmark points.
The higher the value the more penalise the distant location will be. The value 0 means no penalty.}

\item{radius_km}{Numeric. Search radius in kilometers for considering benchmark samples.}

\item{scale_factor}{Numeric. A scaling factor to correct distance calculations by converting degrees to
meters in unprojected coordinate systems. The default is 100,000 for an average conversion in Australia.
On the equator, this factor is approximately 111,235, and it varies with latitude according to a cosine
function.}

\item{k_pred}{Integer. Number of nearest ENV/predicted RS samples to take.}

\item{k_obs}{Integer. Number of nearest observed RS sample to takes.}

\item{bin_width}{Numeric. Specifies the bin width of the histogram. If \code{histogram} is
a \strong{histo} object, this value can be read from its attributes and may be left \code{NULL}.
The bin width must be consistent between the histogram creation and the benchmarking step to
ensure condition is accurately calculated.}

\item{interpolate}{Logical. Should interpolate the histogram?}

\item{offset}{Integer. Specifies the number of histogram bins that were ignored during
normalization (see \code{\link{normalise}}). If \code{histogram} is a \strong{histo} object,
this value can be read from its attributes and may be set \code{NULL}. Similar to bin-width,
the \code{offset} must be consistent between the histogram normalisation and the benchmarking
step to ensure condition is accurately calculated.}

\item{confidence}{Numeric. The confidence value for LDC methods. See details below..}

\item{lambda}{Numeric. The lambda param for LDC Cauchy weighting...}

\item{exclude_slef}{Logical. To exclude a benchmark point from assessing itself.}

\item{drop_features}{Integer vector. Completely remove the RS variable from the benchmarking process. For
consistency, it is recommended to exclude the same variables used in the histogram step; unless
you have a specific reason not to.}

\item{make_su}{Logical. To make the uncertainty map or not.}

\item{num_threads}{Integer. Specifies the number of CPU threads to be used for processing. A value
below 1 indicates that all available threads will be utilized. Refer to the details section for
more information.}

\item{...}{Additional arguments for writing raster outputs e.g. \code{filename},
\code{overwrite}, and \code{wopt} from terra \code{\link[terra]{predict}}.}
}
\value{
A matrix or SpatRaster, depending on the inputs.
}
\description{
The HCAS (Habitat Condition Assessment System) benchmarking function evaluates habitat
condition by comparing observed and predicted remote sensing (RS) variables. It integrates
multiple RS data layers to provide a comprehensive assessment of habitat quality and changes
over time. This function is designed to help researchers and conservationists quantify the
impacts of environmental changes and management interventions on habitat condition.
}
\details{
Ensure that the order of RS variables is consistent between observed and predicted inputs
(for both raster and matrix formats). The RS variable values must be centered and scaled
prior to prediction. Failure to do so may result in variables with larger ranges having
disproportionate influence in the multi-dimensional distance calculations.

Ensure that \href{https://en.wikipedia.org/wiki/OpenMP}{OpenMP} is installed on your system
to take advantage of parallel processing and accelerate computations. While most systems
include OpenMP by default, you may need to load the appropriate module if you're using an HPC
system.

Note that the default parameters are tailored for Australia and may not be suitable for other
regions.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{histogram}}, \code{\link{normalise}}, and \code{\link{calibrate}}
}
