% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate habitat condition output}
\usage{
calibrate(x, x_values, y_values = c(0, 0.101, 0.944, 1), ...)
}
\arguments{
\item{x}{A SpatRaster, matrix, data.frame, or vector containing HCAS habitat condition values from
the benchmarking function (see \code{\link{benchmark}}). If \code{x} is a matrix or data.frame,
the calibration will be applied to all columns.}

\item{x_values}{Numeric vector of uncalibrated condition values. It is recommended that \code{x_values}
cover the full range of values, including the minimum and maximum of the raw condition data.}

\item{y_values}{Numeric vector of calibrated target condition value corresponding to \code{x_values}.}

\item{...}{Additional arguments for writing raster outputs e.g. \code{filename},
\code{overwrite}, and \code{wopt} from terra \code{\link[terra]{predict}}.}
}
\value{
A matrix, SpatRaster or vector, depending on the inputs.
}
\description{
This function calibrates the HCAS habitat condition values and scales them between 0 and 1
using a Spline function.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{benchmark}}
}
