% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate habitat condition output}
\usage{
calibrate(
  x,
  low,
  mid,
  high,
  max,
  target1 = 0.1,
  target2 = 0.944,
  method = "spline",
  ...
)
}
\arguments{
\item{x}{A SpatRaster, matrix, data.frame, or vector containing HCAS habitat condition values from
the benchmarking function (see \code{\link{benchmark}}). If \code{x} is a matrix or data.frame,
the calibration will be applied to all columns.}

\item{low}{Numeric. The median uncalibrated condition value for highly modified areas.}

\item{mid}{Numeric. The median uncalibrated condition value for natural reference areas.}

\item{high}{Numeric. The median uncalibrated condition value for \strong{benchmark} reference
sites used in the \code{\link{benchmark}} function.}

\item{max}{Numeric. The median uncalibrated condition value for the entire area of interest, based
on the maximum values in the raster or matrix.}

\item{target1}{Numeric. The calibrated target value for highly modified areas. The default value of
0.1 is based on Australian data.}

\item{target2}{Numeric. The calibrated target value for natural reference areas. The default value
of 0.944 is based on Australian data.}

\item{method}{Character. Specifies the calibration method, either "linear" or "spline". The spline
method produces a smoother map and histogram of calibrated condition values. The default is "spline".}
}
\value{
A matrix, SpatRaster or vector, depending on the inputs.
}
\description{
This function calibrates the HCAS habitat condition and scales the values between
0 and 1.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{benchmark}}
}
