% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicHCAS.R
\docType{_PACKAGE}
\name{ClassicHCAS}
\alias{ClassicHCAS}
\title{ClassicHCAS: The Classic Habitat Condition Assessment System (HCAS)}
\description{
The HCAS (Habitat Condition Assessment System) benchmarking function evaluates habitat
condition by comparing observed and predicted remote sensing (RS) variables. It integrates
multiple RS data layers to provide a comprehensive assessment of habitat quality and changes
over time. This function is designed to help researchers and conservationists quantify the
impacts of environmental changes and management interventions on habitat condition.
}
\references{
The Habitat Condition Assessment System (HCAS)...
}
\seealso{
\code{\link{benchmark}} and \code{\link{histogram}}.
}
\author{
Roozbeh Valavi, el al. (will be updated)
}
