% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{The HCAS histogram}
\usage{
histogram(
  observed,
  predicted,
  samples_xy,
  radius_km = 1000,
  bin_width = 0.05,
  bin_num = 650,
  num_threads = -1,
  filename = "",
  source_code
)
}
\arguments{
\item{observed}{A matrix or data.frame of observed remote sensing variables.}

\item{predicted}{A matrix or data.frame of predicted remote sensing variables.
Notice that the order or predicted RS variables must the same as observed ones.}

\item{samples_xy}{A matrix or data.frame of x and y (longitude and latitude) of the
reference points used for observed and predicted RS variiables.}

\item{radius_km}{Numeric. Search radius in kilometers for considering reference samples
in creating histogram.}

\item{bin_width}{Numeric. Specifies the bin width of the histogram. Finding the optimal bin width
may require some experimentation to achieve the best results. The bin width is added as an attribute
to the output object, allowing it to be consistently applied in further steps, ensuring accuracy
and consistency in the benchmarking process.}

\item{bin_num}{Integer. Specifies the number of bins for the histogram. It is generally recommended
to use the default value of 650. Adjusting the \code{bin_width} is often more effective than changing
\code{bin_num}.}

\item{num_threads}{Integer. Specifies the number of CPU threads to be used for processing. A value
below 1 indicates that all available threads will be utilized. Refer to the details section for
more information.}

\item{filename}{Char (optional). The output file name for the .text file.}
}
\value{
A histo object (matrix, array)
}
\description{
The HCAS histogram calculation is based on pair-point densities. It is an integral part of HCAS,
designed to learn the expected observed remote sensing (RS) values from the
predicted RS values across a wide range of reference sites. Note that the reference
sites used for the \code{histogram} do not need to be the same as those used in
the \code{\link{benchmark}} function. See more in details.
}
\details{
Ensure that the order of RS variables is consistent between observed and predicted inputs.
The RS variable values must be centered and scaled prior to prediction. Failure to do
so may result in variables with larger ranges having disproportionate influence in the
multi-dimensional distance calculations.

Ensure that \href{https://en.wikipedia.org/wiki/OpenMP}{OpenMP} is installed on your system
to take advantage of parallel processing and accelerate computations. While most systems
include OpenMP by default, you may need to load the appropriate module if you're using an HPC
system.
}
\examples{
\donttest{
library(ClassicHCAS)

obs_dat <- read.csv("")
prd_dat <- read.csv("")


}
}
\seealso{
\code{\link{benchmark}}
}
