% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{The HCAS histogram}
\usage{
histogram(
  observed,
  predicted,
  samples,
  within_km = 1000,
  bin_width = 0.05,
  bin_num = 650,
  num_threads = parallel::detectCores() - 1,
  filename = "",
  source_code
)
}
\arguments{
\item{observed}{The data.frame or matrix of observed remote sensing variables.}

\item{predicted}{The data.frame or matrix of predicted remote sensing variables.}

\item{samples}{Matrix or df. With the XY or XY and RS and ENV values.}

\item{within_km}{Numeric. Search radius in kilometers considering samples.}

\item{bin_width}{Numeric. The bin width of the histogram}

\item{bin_num}{Int. Number of bin for histogram. Keep it at the default 650..}

\item{num_threads}{Int. Number of CPU threads for processing...}

\item{filename}{Char (optional). The output file name for the .text file.}
}
\value{
matrix
}
\description{
The HCAS histogram calculation based on pair-point densities.
}
\examples{
\donttest{
library(ClassicHCAS)

obs_dat <- read.csv("")
prd_dat <- read.csv("")


}
}
\seealso{
\code{\link{benchmark}}
}
