% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Clean and normalise HCAS histogram}
\usage{
normalise(
  x,
  bin_width = NULL,
  trim_size = 400,
  offset = 0,
  legacy = FALSE,
  filename = ""
)
}
\arguments{
\item{x}{An HCAS histo object or a matrix representing the HCAS histogram (see \code{\link{histogram}}).}

\item{bin_width}{Numeric. Specifies the bin width of the histogram. If \code{histogram} is
a \strong{histo} object, this value can be read from its attributes and may be left \code{NULL}.
The bin width must be consistent between the histogram creation and the benchmarking step to
ensure condition is accurately calculated.}

\item{trim_size}{Integer. Defines the number of rows and columns in the trimmed histogram. The default
is 400, and it is generally advisable to retain this default setting.}

\item{offset}{Integer. Specifies the number of histogram bins to ignore during normalization. This value
will be stored as an attribute in the output object.}

\item{legacy}{Logical. Whether to use the legacy C++ code for normalisation (for backward
compatibility) or the modern R version (default). The modern version solves the edge effect
issue without any speed compromise.}

\item{filename}{Char (optional). The output file name for the .text file.}
}
\value{
A histo object (also matrix, array)
}
\description{
This process includes trimming the histogram to remove noise and normalising its values.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{histogram}}, and \code{\link{benchmark}}
}
