% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Clean and normalise HCAS histogram}
\usage{
normalise(x, bin_width = NULL, trim_size = 400, offset = 0, filename = "")
}
\arguments{
\item{x}{An HCAS histo object or a matrix representing the HCAS histogram (see \code{\link{histogram}}).}

\item{bin_width}{Numeric. Specifies the bin width of the histogram. If \code{x} is a
\strong{histo} object, this value can be read from its attributes and may be left \code{NULL}. The bin width
is primarily used to ensure consistency in the benchmarking step by accurately tracking and applying the
bin width used during histogram creation.}

\item{trim_size}{Integer. Defines the number of rows and columns in the trimmed histogram. The default
is 400, and it is generally advisable to retain this default setting.}

\item{offset}{Integer. Specifies the number of histogram bins to ignore during normalization. This value
will be stored as an attribute in the output object.}

\item{filename}{Char (optional). The output file name for the .text file.}
}
\value{
A histo object (matrix, array)
}
\description{
This process includes trimming the histogram to remove noise and normalising its values.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{histogram}}, and \code{\link{benchmark}}
}
