% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Number of samples within a proximity (radial search)}
\usage{
proximity(x, samples_xy, radius_km = 200, num_threads = -1, ...)
}
\arguments{
\item{x}{A SpatRaster representing the study area over which sample density will be calculated.}

\item{samples_xy}{A matrix or data.frame containing x and y coordinates (longitude and latitude)
of the reference points used for density calculation.}

\item{radius_km}{Numeric. Specifies the search radius (buffer) in kilometers.}

\item{num_threads}{Integer. Specifies the number of CPU threads to be used for processing. A value
below 1 indicates that all available threads will be utilized. Refer to the details section for
more information.}

\item{...}{Additional arguments for writing raster outputs e.g. \code{filename},
\code{overwrite}, and \code{wopt} from terra \code{\link[terra]{predict}}.}
}
\value{
A SpatRaster
}
\description{
This function calculates the number of samples (x, y coordinates) within a specified radius
for each pixel in a raster map, using Euclidean distance.
}
\details{
Ensure that \href{https://en.wikipedia.org/wiki/OpenMP}{OpenMP} is installed on your system
to take advantage of parallel processing and accelerate computations. While most systems
include OpenMP by default, you may need to load the appropriate module if you're using an HPC
system.
}
\examples{
\donttest{
library(ClassicHCAS)



}
}
\seealso{
\code{\link{benchmark}}
}
